Moved input validation functions into a file
